Introduced a menu bar for naviagation and sign-ins
enabled sessesions so you can be logged in and then logged out.

did with javascript and not a php embed/index.php because i felt like it.

sign in and sign up now redirect to main page and sign in, with the account name now shown in the top right