All listed items that do not belong the user are now visible on the home page.
this allows you to scroll through without seeing anything you own. this obviosly can be changed but i think it is a good feature.

its derived off the same js loader that loads for the profile, just a different php sctipt that includes the sql query for retreiving data.

Also added functionality that when logged out, forces you to the home page, should hopefully avoid listing items while logged out, and profile issues when logged out. will need a fall back as if the session is removed, it will cause the same effet without hitting the button